Successfully show choices on a poll.
